Prototype page flow:
This commit fills in a skeleton for a possible page flow.

It updates:
1.A bunch of empty html files representing the pages we might
need, and the template directory to place them in.
2.The settings.py file, specifying where the template file is.
This is unfortunately an absolute path, because the internet
made me do it.
3.The urls.py file, allowing us to route to access all those
sweet, sweet empty pages.
4.The views.py file. It includes a bunch of... you guessed it,
empty context, no logic view calls.
5.run.sh, a utility shortcut for firing up the development server.

The _ prefixed html's are meant to be inherited from. _base.html will
be the site-wide base, and the other three will inherit from that.
The non-prefixed htmls inherit from the children of _base.