Games collapsed into matches
We did not want to support a feature directly using games. We decided
that for the goals of our project, it made more sense for matches to
be the main object that handles results. Appraising the cost of adding
games later should we decide to, it still made sense to drop them now.

1. Moved game fields to matche, deleted game table.
2. Added match info field to matches.
3. Added starting team elo fields to matches.
4. Dropped join and exit info from team-players. We decided a players
team info would solely be a product of the team they were on. Shmoozers
get to win. Our project philosophy is to only support what our features
require and keep the app light-weight.